Backport PR #15524: Fix visual tests (#15578)
* Backport PR #15524: Fix visual tests

* Try to make the test terminal snapshot more robust

by mock `tree . -L 2` command

* Workaround `test-` dirs issue by capturing `/data` instead of `/`

* Update Playwright Snapshots

* Run customization docs snapshots in serial

* Ignore all test files in galata tests

* Update Playwright Snapshots

* Disable concurrency in `build:examples` to fix flaky test

This was already use in `build:src`

* Ensure no flaky focus indicator on snapshot

(cherry picked from commit c26a0e7cced6eea64a7f0908a7cded8e4c8874fb)

* Update Playwright Snapshots

* Revert some snapshot updates

* Update Playwright Snapshots

* Try to fix "General › Welcome" documentation snapshot

---------
