+++ .licenserc.yaml
@@ -32,6 +32,7 @@ header:
     - 'dev_mode/style.js'
     - 'examples/federated/example.cert'
     - 'galata/test/jupyterlab/notebooks/'
+    - 'galata/test/*/*/*.txt'
     - 'jupyterlab.desktop'
     - 'jupyterlab/staging'
     - 'packages/codemirror/test/foo.grammar'
