+++ galata/test/documentation/commands.test.ts-snapshots/commandsList-documentation-linux.json
@@ -619,14 +619,6 @@
     "shortcuts": []
   },
   {
-    "id": "documentsearch:toggleSearchInSelection",
-    "label": "Search in Selection",
-    "caption": "",
-    "shortcuts": [
-      "Alt L"
-    ]
-  },
-  {
     "id": "editmenu:clear-all",
     "label": "Clear All",
     "caption": "",
@@ -1915,9 +1907,7 @@
     "id": "notebook:toggle-all-cell-line-numbers",
     "label": "Show Line Numbers",
     "caption": "",
-    "shortcuts": [
-      "Shift L"
-    ]
+    "shortcuts": []
   },
   {
     "id": "notebook:toggle-autoclosing-brackets",
