+++ galata/test/documentation/customization.test.ts
@@ -10,6 +10,9 @@ test.use({
   viewport: { height: 720, width: 1280 }
 });
 
+// Use serial mode to avoid flaky screenshots
+test.describe.configure({ mode: 'serial' });
+
 test.describe('Default', () => {
   test('should use default layout', async ({ page }) => {
     await galata.Mock.freezeContentLastModified(page);
