+++ galata/test/documentation/overview.test.ts-snapshots/interface-tabs-documentation-linux.png
